# apigw/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: Apigw API
    version: 0.0.1
paths:
    /pg/apigw/article/v1:
        get:
            tags:
                - Apigw
            operationId: Apigw_ListArticle
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/apigw.v1.ListArticleReply'
        post:
            tags:
                - Apigw
            operationId: Apigw_CreateArticle
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/apigw.v1.Article'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/apigw.v1.CreateArticleReply'
    /pg/apigw/article/v1/{id}:
        get:
            tags:
                - Apigw
            operationId: Apigw_GetArticle
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: int64
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/apigw.v1.GetArticleReply'
        put:
            tags:
                - Apigw
            operationId: Apigw_UpdateArticle
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: int64
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/apigw.v1.UpdateArticleRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/apigw.v1.UpdateArticleReply'
        delete:
            tags:
                - Apigw
            operationId: Apigw_DeleteArticle
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: int64
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/apigw.v1.DeleteArticleReply'
    /pg/apigw/auth/v1:
        post:
            tags:
                - Apigw
            operationId: Apigw_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/apigw.v1.LoginRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/apigw.v1.ReturnTokenReply'
components:
    schemas:
        apigw.v1.Article:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
                title:
                    type: string
                content:
                    type: string
                createTime:
                    type: integer
                    format: int64
        apigw.v1.CreateArticleReply:
            type: object
            properties:
                aid:
                    type: integer
                    format: int64
        apigw.v1.DeleteArticleReply:
            type: object
            properties: {}
        apigw.v1.GetArticleReply:
            type: object
            properties:
                Article:
                    $ref: '#/components/schemas/apigw.v1.Article'
        apigw.v1.ListArticleReply:
            type: object
            properties:
                results:
                    type: array
                    items:
                        $ref: '#/components/schemas/apigw.v1.Article'
        apigw.v1.LoginRequest:
            type: object
            properties:
                username:
                    type: string
                password:
                    type: string
        apigw.v1.RegisterRequest:
            type: object
            properties:
                name:
                    type: string
                password:
                    type: string
                email:
                    type: string
                avatar:
                    type: string
        apigw.v1.ReturnTokenReply:
            type: object
            properties:
                token:
                    type: string
        apigw.v1.UpdateArticleReply:
            type: object
            properties:
                Article:
                    $ref: '#/components/schemas/apigw.v1.Article'
        apigw.v1.UpdateArticleRequest:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
                title:
                    type: string
                content:
                    type: string
tags:
    - name: Apigw